chore: refactored and moved out common test code to root test directory
